Move BarScale component to components-styled
