+++ packages/app/src/components-styled/bar-scale.tsx
